[issue/64] - added method to send database, switched code in lib/api
